Panopticon registration
As a developer
I want to be able to register all finders with panopticon
So that they appear on the main GOV.UK domain

Adds a rake task to do this registration. The idea is to invoke this
rake task whenever the finders are deployed. The definition of finders
will be in json schemas contained within this app.