Remove Documents and Schema Registering
Now that we're using Rummager for returning the documents in a Finder,
we no longer need any of the tests related to any of this code. Also,
there is an issue with schemas being updated so we have now removed
registering the schemas, as they are now just used to serve the list of
possible facets for the Finder filters.