Allow Finders to be published to the PublishingAPI
With the move of GOV.UK to use the Content Store, this commit allows us
to now publish Finders to the PublishingAPI. It contains a
PublishingAPIPublisher class which handles exporting the Finder and it's
signup page, a ContentItemPresenter class which has 2 child classes for
the Finder and the Signup page. It also contains a spec for the
PublishingAPIPublisher class and a rake task to trigger publishing.