Use format instead of slug for email lists
Previously we were using the finder’s slug to set the format of an
email subscription list. This was incorrect as there is differences
between the slugs and formats. (underscores/hyphens, plural/singular
and in the case of drug safety updates completely different)

This commit adds the formats to the metadata json files, and updated
the email list creator to use them instead of slugs.