Move drug device alerts to new content format
By no longer supplying a key in the email_signup_choices hash, and
making it an array, we can simplify a lot of our code.

This commit also renames ‘name’ to ‘radio_button_name’, to distinguish
it from the ‘topic_name’, which has been added. This is due to
sometimes wanting to display something different in the topic title of
a subscription from what was on the radio button.